+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the "AddFolderToWorkspace" extension will be documented in this file.
 
+## [1.1.1]
+
+`AddFolderToWorkspace: Add manually a directory...` - Paths can be added manually from now on. After that, the new workspace can be saved directly in the settings.s
+
+Fixed bug - when a workspace has already been added.
+
 ## [1.1.0]
 
 ### AddFolderToWorkspace